финальное задание спринт 12

+++ src/main/java/ru/yandex/practicum/filmorate/storage/interfaces/UserStorage.java
@@ -12,6 +12,6 @@ public interface UserStorage {
 
     Collection<User> getAllUsers();
 
-    void delete(long id);
+    void delete(int id);
 
 }
